Refactored FileParserUtil
